proyectoheal/requirements.txt
azure==1.0.3
azure-common==1.1.2
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==1.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==1.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
bzr==2.6b1
colorama==0.2.5
configobj==4.7.2
cycler==0.10.0
decorator==3.4.0
django==1.9.4
django-crispy-forms==1.6.0
django-datetime-widget==0.9.3
django-registration==2.0.4
django-registration-redux==1.4
futures==3.0.5
html5lib==0.999
httplib2==0.8
ipython==1.2.1
keyring==3.5
launchpadlib==1.10.2
lazr.authentication==0.1.3
lazr.restfulclient==0.13.1
lazr.uri==1.0.3
lockfile==0.8
lpthw.web==1.1
matplotlib==1.5.1
mercurial==3.7.3
MySQL-python==1.2.4b4
numpy==1.11.0
oauth==1.0.1
paramiko==1.10.1
pbr==1.9.1
pexpect==4.0.1
Pillow==3.1.1
psycopg2==2.6.1
ptyprocess==0.5.1
pycrypto==2.6.1
pyparsing==2.0.1
pyreadline==2.1
python-dateutil==2.5.2
pytz==2012rc0
requests==2.2.1
scipy==0.16.1
SecretStorage==2.0.0
simplegeneric==0.8.1
simplejson==3.3.1
six==1.5.2
stevedore==0.14.1
testresources==2.0.0
urllib3==1.7.1
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wadllib==1.3.2
wsgi-intercept==1.2.2
zope.interface==4.0.5
